Port Prio2 to a VDAF
Defines a new `Vdaf` implementor that wraps the ENPA Prio system into a
VDAF. The implementation is backwards compatible with `Client` and
`Server`. Method `Prio2::prepare_init_with_query_rand()` can be used as
an alternative to the `Aggregator` implementation for applicaitons, like
ENPA, where the query randomness is picked by a third-party.